+++ CatBot/__init__.py
@@ -6,4 +6,4 @@ CatBot
 Package containing all the source code for CatBot.
 """
 
-from . import internal, representations
+from . import internal